Replaced `ThunderstoreClient` with `HttpApiClient`
